requirements.txt
fastapi==0.111.0
keras==3.2.1
numpy==1.26.4
Pillow==10.3.0
prometheus_client==0.20.0
tensorflow==2.16.1
uvicorn==0.29.0
python-multipart
tensorflow
